TestControllerを新規作成し、GETおよびPOSTメソッドのテスト用エンドポイントを追加
